module7/mwe/vector-retrieval/requirements.txt
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy>=1.24.0